requirements.txt
streamlit
requests
SpeechRecognition
PyAudio
pyttsx3
python-dotenv
google-generativeai